requirements.txt
google-generativeai
transformers
torch
huggingface_hub
pandas 